+++ src/index.ts
@@ -1,3 +1,3 @@
-export type Namespace = string | string[]
+export type Namespace = string
 export { default as ConnectedProvider } from './ConnectedProvider'
 export { default as useConnectedReducer } from './useConnectedReducer'