+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - **BREAKING**: States are now stored in a provider component. Every component that wants to use connected reducers has to be an ancestor of this `ConnectedProvider`. This provides an easy way to debug connected states via React Dev Tools.
+- **BREAKING**: Namespaces are not allowed to be an array anymore, but have to be strings.
 
 ## [0.2.0] - 2019-03-25
 ### Changed